requirements.txt
pytorch-lightning==2.5.*
numpy==2.3.*
pandas==2.3.*
wbgapi==1.0.12
scikit-learn==1.7.*
h5py==3.14.0
ipykernel==6.29.5
statsmodels==0.14.*
tables==3.10.2